requirements.txt
flask==3.0.0
werkzeug==3.0.1
python-frontmatter==1.0.0
beautifulsoup4==4.12.2
markdown==3.5.1
openai==1.3.5
python-dotenv==1.0.0
tqdm==4.66.1
jinja2==3.1.2
PyYAML==6.0.1 